.history/requirements_20240917154843.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.1.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
et-xmlfile==1.1.0
exceptiongroup==1.2.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
numpy==2.0.1
openpyxl==3.1.5
pandas==2.2.2
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.4
pytz==2024.1
requests==2.32.3
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
selenium

